# Avoid allocations by inlining functions accepting accumulators