--- a/src/accumulators/compSum.jl
+++ b/src/accumulators/compSum.jl
@@ -3,24 +3,24 @@ mutable struct CompSumAcc{T, EFT}
     e :: T
 end
 
-compSumAcc(EFT) = T->compSumAcc(EFT, T)
-compSumAcc(EFT, T) = CompSumAcc{T, EFT}(vzero(T), vzero(T))
+@inline compSumAcc(EFT) = T->compSumAcc(EFT, T)
+@inline compSumAcc(EFT, T) = CompSumAcc{T, EFT}(vzero(T), vzero(T))
 
-function add!(acc::CompSumAcc{T, EFT}, x::T) where {T, EFT}
+@inline function add!(acc::CompSumAcc{T, EFT}, x::T) where {T, EFT}
     SIMDops.@explicit
 
     acc.s, e = EFT(acc.s, x)
     acc.e += e
 end
 
-function add!(acc::A, x::A) where {A<:CompSumAcc{T, EFT}} where {T, EFT}
+@inline function add!(acc::A, x::A) where {A<:CompSumAcc{T, EFT}} where {T, EFT}
     SIMDops.@explicit
 
     acc.s, e = EFT(acc.s, x.s)
     acc.e += x.e + e
 end
 
-function Base.sum(acc::CompSumAcc{T, EFT}) where {T<:Vec, EFT}
+@inline function Base.sum(acc::CompSumAcc{T, EFT}) where {T<:Vec, EFT}
     acc_r = compSumAcc(EFT, fptype(T))
     acc_r.e = vsum(acc.e)
     for xi in acc.s
@@ -30,6 +30,6 @@ function Base.sum(acc::CompSumAcc{T, EFT}) where {T<:Vec, EFT}
     acc_r
 end
 
-function Base.sum(acc::CompSumAcc{T, EFT}) where {T<:Real, EFT}
+@inline function Base.sum(acc::CompSumAcc{T, EFT}) where {T<:Real, EFT}
     acc.s + acc.e
 end